third_party/blink/web_tests/external/wpt/streams/readable-streams/global-expected.txt

This is a testharness.js-based test.
[FAIL] Cancel promise is created in same global as stream
  assert_equals: Cancel promise comes from the same global as the stream expected true but got false
[FAIL] Reader objects in correct global
  assert_equals: Object comes from other realm expected true but got false
Harness: the test ran to completion.
